+++ debug/plugins/apexcharts/src/svgjs/svg.js
@@ -1135,6 +1135,18 @@
 
 
 
+
+
+
+
+
+
+
+
+
+
+
+
         // functionToCall: [list of morphable objects]
         // e.g. move: [SVG.Number, SVG.Number]
       };this.attrs = {
@@ -1192,24 +1204,12 @@
           this.situation = this.situations.shift();if (this.situation) {if (this.situation instanceof SVG.Situation) {this.start();} else {// If it is not a SVG.Situation, then it is a function, we execute it
               this.situation.call(this);}}return this;}, // updates all animations to the current state of the element
         // this is important when one property could be changed from another property
-        initAnimations: function () {var source;var s = this.situation;if (s.init) return this;for (var i in s.animations) {source = this.target()[i]();if (!Array.isArray(source)) {source = [source];}if (!Array.isArray(s.animations[i])) {s.animations[i] = [s.animations[i]];}
-
-            // if(s.animations[i].length > source.length) {
+        initAnimations: function () {var source;var s = this.situation;if (s.init) return this;for (var i in s.animations) {source = this.target()[i]();if (!Array.isArray(source)) {source = [source];}if (!Array.isArray(s.animations[i])) {s.animations[i] = [s.animations[i]];} // if(s.animations[i].length > source.length) {
             //  source.concat = source.concat(s.animations[i].slice(source.length, s.animations[i].length))
             // }
-
-            for (var j = source.length; j--;) {
-              // The condition is because some methods return a normal number instead
+            for (var j = source.length; j--;) {// The condition is because some methods return a normal number instead
               // of a SVG.Number
-              if (s.animations[i][j] instanceof SVG.Number) {source[j] = new SVG.Number(source[j]);}
-
-              s.animations[i][j] = source[j].morph(s.animations[i][j]);
-            }
-          }
-
-          for (var i in s.attrs) {
-            s.attrs[i] = new SVG.MorphObj(this.target().attr(i), s.attrs[i]);
-          }
+              if (s.animations[i][j] instanceof SVG.Number) {source[j] = new SVG.Number(source[j]);}s.animations[i][j] = source[j].morph(s.animations[i][j]);}}for (var i in s.attrs) {s.attrs[i] = new SVG.MorphObj(this.target().attr(i), s.attrs[i]);}
 
           for (var i in s.styles) {
             s.styles[i] = new SVG.MorphObj(this.target().style(i), s.styles[i]);
@@ -1794,6 +1794,24 @@
 
 
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
             // the element is NOT in the dom, throw error
             // disabling the check below which fixes issue #76
             // if (!document.documentElement.contains(element.node)) throw new Exception('Element not in the dom')
@@ -1842,30 +1860,12 @@
       point: function (x, y) {return new SVG.Point(x, y).transform(this.screenCTM().inverse());} });SVG.extend(SVG.Element, { // Set svg element attribute
       attr: function (a, v, n) {// act as full getter
         if (a == null) {// get an object of attributes
-          a = {};v = this.node.attributes;for (var n = v.length - 1; n >= 0; n--) {a[v[n].nodeName] = SVG.regex.isNumber.test(v[n].nodeValue) ? parseFloat(v[n].nodeValue) : v[n].nodeValue;}return a;
-        } else if (typeof a === 'object') {
-          // apply every attribute individually if an object is passed
-          for (var v_ in a) this.attr(v_, a[v_]);
-        } else if (v === null) {
-          // remove value
-          this.node.removeAttribute(a);
-        } else if (v == null) {
-          // act as a getter if the first and only argument is not an object
-          v = this.node.getAttribute(a);
-          return v == null ?
-          SVG.defaults.attrs[a] :
-          SVG.regex.isNumber.test(v) ?
-          parseFloat(v) : v;
-        } else {
-          // BUG FIX: some browsers will render a stroke if a color is given even though stroke width is 0
-          if (a == 'stroke-width') {this.attr('stroke', parseFloat(v) > 0 ? this._stroke : null);} else if (a == 'stroke') {this._stroke = v;}
-
-          // convert image fill and stroke to patterns
-          if (a == 'fill' || a == 'stroke') {
-            if (SVG.regex.isImage.test(v)) {v = this.doc().defs().image(v, 0, 0);}
-
-            if (v instanceof SVG.Image) {
-              v = this.doc().defs().pattern(0, 0, function () {
+          a = {};v = this.node.attributes;for (var n = v.length - 1; n >= 0; n--) {a[v[n].nodeName] = SVG.regex.isNumber.test(v[n].nodeValue) ? parseFloat(v[n].nodeValue) : v[n].nodeValue;}return a;} else if (typeof a === 'object') {// apply every attribute individually if an object is passed
+          for (var v_ in a) this.attr(v_, a[v_]);} else if (v === null) {// remove value
+          this.node.removeAttribute(a);} else if (v == null) {// act as a getter if the first and only argument is not an object
+          v = this.node.getAttribute(a);return v == null ? SVG.defaults.attrs[a] : SVG.regex.isNumber.test(v) ? parseFloat(v) : v;} else {// BUG FIX: some browsers will render a stroke if a color is given even though stroke width is 0
+          if (a == 'stroke-width') {this.attr('stroke', parseFloat(v) > 0 ? this._stroke : null);} else if (a == 'stroke') {this._stroke = v;} // convert image fill and stroke to patterns
+          if (a == 'fill' || a == 'stroke') {if (SVG.regex.isImage.test(v)) {v = this.doc().defs().image(v, 0, 0);}if (v instanceof SVG.Image) {v = this.doc().defs().pattern(0, 0, function () {
                 this.add(v);
               });
             }
@@ -2516,6 +2516,12 @@
 
 
 
+
+
+
+
+
+
     // Get all siblings, including myself
   });SVG.Gradient = SVG.invent({ // Initialize node
       create: function (type) {this.constructor.call(this, SVG.create(type + 'Gradient')); // store type
@@ -2535,35 +2541,29 @@
       from: function (x, y) {return (this._target || this).type == 'radial' ? this.attr({ fx: new SVG.Number(x), fy: new SVG.Number(y) }) : this.attr({ x1: new SVG.Number(x), y1: new SVG.Number(y) });}, // To position
       to: function (x, y) {return (this._target || this).type == 'radial' ? this.attr({ cx: new SVG.Number(x), cy: new SVG.Number(y) }) : this.attr({ x2: new SVG.Number(x), y2: new SVG.Number(y) });} }); // Base gradient generation
   SVG.extend(SVG.Defs, { // define gradient
-      gradient: function (type, block) {return this.put(new SVG.Gradient(type)).update(block);} });SVG.Stop = SVG.invent({
-    // Initialize node
-    create: 'stop',
+      gradient: function (type, block) {return this.put(new SVG.Gradient(type)).update(block);} });SVG.Stop = SVG.invent({ // Initialize node
+      create: 'stop', // Inherit from
+      inherit: SVG.Element, // Add class methods
+      extend: { // add color stops
+        update: function (o) {
+          if (typeof o === 'number' || o instanceof SVG.Number) {
+            o = {
+              offset: arguments[0],
+              color: arguments[1],
+              opacity: arguments[2]
+            };
+          }
 
-    // Inherit from
-    inherit: SVG.Element,
+          // set attributes
+          if (o.opacity != null) this.attr('stop-opacity', o.opacity);
+          if (o.color != null) this.attr('stop-color', o.color);
+          if (o.offset != null) this.attr('offset', new SVG.Number(o.offset));
 
-    // Add class methods
-    extend: {
-      // add color stops
-      update: function (o) {
-        if (typeof o === 'number' || o instanceof SVG.Number) {
-          o = {
-            offset: arguments[0],
-            color: arguments[1],
-            opacity: arguments[2]
-          };
+          return this;
         }
-
-        // set attributes
-        if (o.opacity != null) this.attr('stop-opacity', o.opacity);
-        if (o.color != null) this.attr('stop-color', o.color);
-        if (o.offset != null) this.attr('offset', new SVG.Number(o.offset));
-
-        return this;
       }
-    }
 
-  });
+    });
 
   SVG.Pattern = SVG.invent({
     // Initialize node