+++ global.d.ts
@@ -1,7 +1,6 @@
 interface Window {
     observable: Observable;
     Observer : Observer;
-    R : R;
     flutter_inappwebview : flutter_inappwebview;
     cacheStorage: CacheLocalStorage;
     cacheMemory: CacheMemory;