+++ install.sh
@@ -90,7 +90,6 @@ module.exports = {
             ,"@flexvue/fastrouter":`./${debugPath}/flexvue/core/fastrouter.class.js`
             ,"@flexvue/recyclerview":`./${debugPath}/flexvue/core/recyclerview.class.js`
             ,"@flexvue/caches":`./${debugPath}/flexvue/core/caches.class.js`
-            ,"@flexvue/resource":`./${debugPath}/flexvue/core/resource.class.js`
             ,"@editorjs/editorjs": `./${debugPath}/plugins/@editorjs/editorjs/dist/editorjs.mjs`
             ,"@editorjs/simple-image": `./${debugPath}/plugins/@editorjs/simple-image/dist/simple-image.mjs`
             ,"@editorjs/header": `./${debugPath}/plugins/@editorjs/header/dist/header.mjs`
@@ -156,7 +155,6 @@ echo '{
           "@flexvue/fastrouter": ["src/flexvue/core/fastrouter.class.ts"],
           "@flexvue/recyclerview": ["src/flexvue/core/recyclerview.class.ts"],
           "@flexvue/caches": ["src/flexvue/core/caches.class.ts"],
-          "@flexvue/resource": ["src/flexvue/core/resource.class.ts"],
           "@editorjs/editorjs": ["node_modules/@editorjs/editorjs"],
           "@editorjs/simple-image": ["node_modules/@editorjs/simple-image"],
           "@editorjs/header": ["node_modules/@editorjs/header"],