+++ src/example/r/r.ts
@@ -1,10 +1,8 @@
-import R from '@flexvue/resource';
-
 const onReady = () : void =>
 {
     // R 클래스를 import하고, sysmsg 리소스에 접근하여 사용할 수 있습니다.
     // Log.d(window.R.sysmsg); // sysmsg 리소스 데이터 출력
-    Log.d(window.R.arrays); // arrays 리소스 데이터 출력
+    Log.d(R.arrays); // arrays 리소스 데이터 출력
     // Log.d(window.R.strings); // strings 리소스 데이터 출력
     // Log.d(window.R.numbers); // numbers 리소스 데이터 출력
 
@@ -15,8 +13,7 @@ document.addEventListener("DOMContentLoaded", () => {
     config.surport_langs = ['en'];
 
     // R 클래스 초기화 후에 DOMContentLoaded 이벤트 발생
-    window.R = R;
-    window.R.__init({
+    R.__init({
         // sysmsg : new URL(`../js/values/sysmsg${App.getLocale()}.js`, import.meta.url).href,
         arrays : new URL(`../js/values/arrays${App.getLocale()}.js`, import.meta.url).href,
         // strings: new URL(`../js/values/strings${App.getLocale()}.js`, import.meta.url).href,