+++ src/t1/manager/manager.class.ts
@@ -65,11 +65,11 @@ class ComponentActivity
                     }
                 });
                 if(chks.length < 1){
-                    alert(window.R.sysmsg.i_checkbox);
+                    alert(R.sysmsg.i_checkbox);
                     return false;
                 }
 
-                let cf = confirm(window.R.sysmsg.i_confirm_delete);
+                let cf = confirm(R.sysmsg.i_confirm_delete);
                 if(cf)
                 {
                     ProgressBars.show();